Move InvocationRateLimiter to a separate file
