+++ extensions/ql-vscode/test/vscode-tests/no-workspace/helpers.test.ts
@@ -1,17 +1,4 @@
-import {
-  EnvironmentVariableCollection,
-  EnvironmentVariableMutator,
-  Event,
-  ExtensionContext,
-  ExtensionMode,
-  Memento,
-  SecretStorage,
-  SecretStorageChangeEvent,
-  Uri,
-  window,
-  workspace,
-  WorkspaceFolder,
-} from "vscode";
+import { Uri, window, workspace, WorkspaceFolder } from "vscode";
 import { dump } from "js-yaml";
 import * as tmp from "tmp";
 import { join } from "path";
@@ -28,7 +15,6 @@ import { DirResult } from "tmp";
 import {
   getFirstWorkspaceFolder,
   getInitialQueryContents,
-  InvocationRateLimiter,
   isFolderAlreadyInWorkspace,
   isLikelyDatabaseRoot,
   isLikelyDbLanguageFolder,
@@ -45,118 +31,6 @@ import { Setting } from "../../../src/config";
 import { createMockCommandManager } from "../../__mocks__/commandsMock";
 
 describe("helpers", () => {
-  describe("Invocation rate limiter", () => {
-    // 1 January 2020
-    let currentUnixTime = 1577836800;
-
-    function createDate(dateString?: string): Date {
-      if (dateString) {
-        return new Date(dateString);
-      }
-      const numMillisecondsPerSecond = 1000;
-      return new Date(currentUnixTime * numMillisecondsPerSecond);
-    }
-
-    function createInvocationRateLimiter<T>(
-      funcIdentifier: string,
-      func: () => Promise<T>,
-    ): InvocationRateLimiter<T> {
-      return new InvocationRateLimiter(
-        new MockExtensionContext(),
-        funcIdentifier,
-        func,
-        (s) => createDate(s),
-      );
-    }
-
-    it("initially invokes function", async () => {
-      let numTimesFuncCalled = 0;
-      const invocationRateLimiter = createInvocationRateLimiter(
-        "funcid",
-        async () => {
-          numTimesFuncCalled++;
-        },
-      );
-      await invocationRateLimiter.invokeFunctionIfIntervalElapsed(100);
-      expect(numTimesFuncCalled).toBe(1);
-    });
-
-    it("doesn't invoke function again if no time has passed", async () => {
-      let numTimesFuncCalled = 0;
-      const invocationRateLimiter = createInvocationRateLimiter(
-        "funcid",
-        async () => {
-          numTimesFuncCalled++;
-        },
-      );
-      await invocationRateLimiter.invokeFunctionIfIntervalElapsed(100);
-      await invocationRateLimiter.invokeFunctionIfIntervalElapsed(100);
-      expect(numTimesFuncCalled).toBe(1);
-    });
-
-    it("doesn't invoke function again if requested time since last invocation hasn't passed", async () => {
-      let numTimesFuncCalled = 0;
-      const invocationRateLimiter = createInvocationRateLimiter(
-        "funcid",
-        async () => {
-          numTimesFuncCalled++;
-        },
-      );
-      await invocationRateLimiter.invokeFunctionIfIntervalElapsed(100);
-      currentUnixTime += 1;
-      await invocationRateLimiter.invokeFunctionIfIntervalElapsed(2);
-      expect(numTimesFuncCalled).toBe(1);
-    });
-
-    it("invokes function again immediately if requested time since last invocation is 0 seconds", async () => {
-      let numTimesFuncCalled = 0;
-      const invocationRateLimiter = createInvocationRateLimiter(
-        "funcid",
-        async () => {
-          numTimesFuncCalled++;
-        },
-      );
-      await invocationRateLimiter.invokeFunctionIfIntervalElapsed(0);
-      await invocationRateLimiter.invokeFunctionIfIntervalElapsed(0);
-      expect(numTimesFuncCalled).toBe(2);
-    });
-
-    it("invokes function again after requested time since last invocation has elapsed", async () => {
-      let numTimesFuncCalled = 0;
-      const invocationRateLimiter = createInvocationRateLimiter(
-        "funcid",
-        async () => {
-          numTimesFuncCalled++;
-        },
-      );
-      await invocationRateLimiter.invokeFunctionIfIntervalElapsed(1);
-      currentUnixTime += 1;
-      await invocationRateLimiter.invokeFunctionIfIntervalElapsed(1);
-      expect(numTimesFuncCalled).toBe(2);
-    });
-
-    it("invokes functions with different rate limiters", async () => {
-      let numTimesFuncACalled = 0;
-      const invocationRateLimiterA = createInvocationRateLimiter(
-        "funcid",
-        async () => {
-          numTimesFuncACalled++;
-        },
-      );
-      let numTimesFuncBCalled = 0;
-      const invocationRateLimiterB = createInvocationRateLimiter(
-        "funcid",
-        async () => {
-          numTimesFuncBCalled++;
-        },
-      );
-      await invocationRateLimiterA.invokeFunctionIfIntervalElapsed(100);
-      await invocationRateLimiterB.invokeFunctionIfIntervalElapsed(100);
-      expect(numTimesFuncACalled).toBe(1);
-      expect(numTimesFuncBCalled).toBe(1);
-    });
-  });
-
   describe("codeql-database.yml tests", () => {
     let dir: tmp.DirResult;
     let language: QueryLanguage;
@@ -249,116 +123,6 @@ describe("helpers", () => {
       expect(await isLikelyDbLanguageFolder(nestedDbPythonFolder)).toBe(true);
     });
   });
-
-  class MockExtensionContext implements ExtensionContext {
-    extensionMode: ExtensionMode = 3;
-    subscriptions: Array<{ dispose(): unknown }> = [];
-    workspaceState: Memento = new MockMemento();
-    globalState = new MockGlobalStorage();
-    extensionPath = "";
-    asAbsolutePath(_relativePath: string): string {
-      throw new Error("Method not implemented.");
-    }
-    storagePath = "";
-    globalStoragePath = "";
-    logPath = "";
-    extensionUri = Uri.parse("");
-    environmentVariableCollection = new MockEnvironmentVariableCollection();
-    secrets = new MockSecretStorage();
-    storageUri = Uri.parse("");
-    globalStorageUri = Uri.parse("");
-    logUri = Uri.parse("");
-    extension: any;
-  }
-
-  class MockEnvironmentVariableCollection
-    implements EnvironmentVariableCollection
-  {
-    [Symbol.iterator](): Iterator<
-      [variable: string, mutator: EnvironmentVariableMutator],
-      any,
-      undefined
-    > {
-      throw new Error("Method not implemented.");
-    }
-    persistent = false;
-    replace(_variable: string, _value: string): void {
-      throw new Error("Method not implemented.");
-    }
-    append(_variable: string, _value: string): void {
-      throw new Error("Method not implemented.");
-    }
-    prepend(_variable: string, _value: string): void {
-      throw new Error("Method not implemented.");
-    }
-    get(_variable: string): EnvironmentVariableMutator | undefined {
-      throw new Error("Method not implemented.");
-    }
-    forEach(
-      _callback: (
-        variable: string,
-        mutator: EnvironmentVariableMutator,
-        collection: EnvironmentVariableCollection,
-      ) => any,
-      _thisArg?: any,
-    ): void {
-      throw new Error("Method not implemented.");
-    }
-    delete(_variable: string): void {
-      throw new Error("Method not implemented.");
-    }
-    clear(): void {
-      throw new Error("Method not implemented.");
-    }
-  }
-
-  class MockMemento implements Memento {
-    keys(): readonly string[] {
-      throw new Error("Method not implemented.");
-    }
-    map = new Map<any, any>();
-
-    /**
-     * Return a value.
-     *
-     * @param key A string.
-     * @param defaultValue A value that should be returned when there is no
-     * value (`undefined`) with the given key.
-     * @return The stored value or the defaultValue.
-     */
-    get<T>(key: string, defaultValue?: T): T {
-      return this.map.has(key) ? this.map.get(key) : defaultValue;
-    }
-
-    /**
-     * Store a value. The value must be JSON-stringifyable.
-     *
-     * @param key A string.
-     * @param value A value. MUST not contain cyclic references.
-     */
-    async update(key: string, value: any): Promise<void> {
-      this.map.set(key, value);
-    }
-  }
-
-  class MockGlobalStorage extends MockMemento {
-    public setKeysForSync(_keys: string[]): void {
-      return;
-    }
-  }
-
-  class MockSecretStorage implements SecretStorage {
-    get(_key: string): Thenable<string | undefined> {
-      throw new Error("Method not implemented.");
-    }
-    store(_key: string, _value: string): Thenable<void> {
-      throw new Error("Method not implemented.");
-    }
-    delete(_key: string): Thenable<void> {
-      throw new Error("Method not implemented.");
-    }
-    onDidChange!: Event<SecretStorageChangeEvent>;
-  }
 
   it("should report stream progress", () => {
     const progressSpy = jest.fn();
